Move DatadogClassloader to bootstrap
